requirements.txt
arrow==1.3.0
beautifulsoup4==4.13.3
binaryornot==0.4.4
botcity-email-plugin==0.1.5
botcity-framework-base==0.4.4
botcity-framework-web==0.10.0
botcity-gmail-plugin==0.2.5
botcity-maestro-sdk==0.6.0
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
distro==1.9.0
dotenv==0.9.9
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
httplib2==0.22.0
idna==3.10
imap-tools==1.10.0
importlib_metadata==8.6.1
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msedge-selenium-tools==3.141.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
selenium==3.141.0
six==1.17.0
soupsieve==2.6
text-unidecode==1.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
webdriver-manager==4.0.2
zipp==3.21.0
